LayoutTests/webaudio/audiochannelsplitter-expected.txt

Tests AudioChannelSplitter.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS Exception been thrown for numberOfOutputs <= 0.
PASS Exception been thrown for numberOfOutputs >= 32.
PASS AudioChannelSplitter created successfully with numberOfOutputs = 32.
PASS AudioChannelSplitter has 32 outputs when it is created with numberOfOutputs = 32.
PASS AudioChannelSplitter has one input.
PASS AudioChannelSplitter created successfully with empty parameter.
PASS AudioChannelSplitter has 6 outputs when it is created with empty parameter.
PASS Correctly exchanged left and right channels.
PASS successfullyParsed is true

TEST COMPLETE
